refactor: logout 버튼 logic 코드 이동
